feat: replace int to float

+++ Exercises/04.py
@@ -3,7 +3,7 @@ print('<<Salary Increase>>');
 print();
 salary = float(input('Enter your salary: '));
 print();
-percent = int(input('Enter the percentage increase: '));
+percent = float(input('Enter the percentage increase: '));
 increase = (percent * salary) / 100;
 print();
 print(f'RESULT: With the {percent}% increase (+{increase:.2f}), your salary went from {salary:.2f} to {(salary + increase):.2f}');
